+++ config/splat.dra.yaml
@@ -63,6 +63,7 @@ segments:
       - [0x41F70, rodata]
       - [0x41FB0, rodata]
       - [0x42000, rodata]
+      - [0x42354, .rodata, 953A0]
       - [0x42398, c, 42398]
       - [0x46FD4, c, 46FD4]
       - [0x4768C, c, 4768C]
@@ -72,5 +73,7 @@ segments:
       - [0x73D7C, c, 73D7C]
       - [0x75DA0, c, 75DA0]
       - [0x75F54, c, 75F54]
+      - [0x953A0, c, 953A0]
+      - [0x95484, c, 95484]
       - [0x962AC, data]
   - [0x119870]
