+++ src/dra/953A0.c
@@ -1,0 +1,41 @@
+#include "common.h"
+#include "dra.h"
+#include "objects.h"
+#include "sfx.h"
+
+void func_801353A0(void) {
+    if (D_801396F4 == 0)
+        return;
+
+    switch (D_80139868[0]) {
+    case 2:
+        func_80133604();
+        break;
+    case 3:
+        func_80133488();
+        break;
+    case 4:
+        func_80132F60();
+        break;
+    case 6:
+        func_80133290();
+        break;
+    case 8:
+        func_80134388();
+        break;
+    case 10:
+        func_801341B4();
+        break;
+    case 12:
+        func_80133960();
+        break;
+    case 14:
+        func_80133BDC();
+        break;
+    case 16:
+        D_8013B684 = 0;
+    case 0:
+        func_80132E38();
+        break;
+    }
+}
